refactor(items): rename change notice UI directories and component files
