+++ .ai/plans/2026-07-28-change-notice-code-rename.md
@@ -45,7 +45,7 @@ not by the compiler.
 - [x] Task 1: Create the branch and record the pre-rename baseline
 - [x] Task 2: Rename the models layer (`items.models.ts`)
 - [x] Task 3: Rename the service + server layer
-- [ ] Task 4: Rename the UI directories and component files
+- [x] Task 4: Rename the UI directories and component files
 - [ ] Task 5: Rename component identifiers, props and local variables
 - [ ] Task 6: Rename route file-internal identifiers and remaining call sites
 - [ ] Task 7: Audit the do-not-rename list
