+++ apps/erp/app/modules/items/ui/ChangeNotice/ImpactPanel.tsx
