+++ apps/erp/app/modules/items/ui/ChangeNotice/affectedItem.types.ts
