+++ apps/erp/app/modules/items/ui/ChangeNoticeActions/index.ts
@@ -1,0 +1,2 @@
+export { default as ChangeOrderRequiredActionForm } from "./ChangeNoticeRequiredActionForm";
+export { default as ChangeOrderRequiredActionsTable } from "./ChangeNoticeRequiredActionsTable";
