+++ apps/erp/app/modules/items/ui/ChangeNoticeTypes/index.ts
@@ -1,0 +1,2 @@
+export { default as ChangeOrderTypeForm } from "./ChangeNoticeTypeForm";
+export { default as ChangeOrderTypesTable } from "./ChangeNoticeTypesTable";
