--- apps/erp/app/modules/items/ui/ChangeOrder/index.ts
@@ -1,19 +1,0 @@
-export { default as AffectedItemDetail } from "./AffectedItemDetail";
-export type * from "./affectedItem.types";
-export { default as ChangeOrderActions } from "./ChangeOrderActions";
-export { default as ChangeOrderChanges } from "./ChangeOrderChanges";
-export * from "./ChangeOrderContent";
-export { default as ChangeOrderDiffViewer } from "./ChangeOrderDiffViewer";
-export { default as ChangeOrderExplorer } from "./ChangeOrderExplorer";
-export { default as ChangeOrderForm } from "./ChangeOrderForm";
-export { default as ChangeOrderHeader } from "./ChangeOrderHeader";
-export { default as ChangeOrderProperties } from "./ChangeOrderProperties";
-export { default as ChangeOrderReleaseMerge } from "./ChangeOrderReleaseMerge";
-export { default as ChangeOrderStatus } from "./ChangeOrderStatus";
-export { default as ChangeOrdersTable } from "./ChangeOrdersTable";
-export { default as CreateChangeOrderModal } from "./CreateChangeOrderModal";
-export type { ChangeOrderImpactItem } from "./ImpactPanel";
-export { default as ImpactPanel } from "./ImpactPanel";
-export * from "./ItemChangeOrderLock";
-export { default as ItemChangeOrders } from "./ItemChangeOrders";
-export { default as ItemOpenChangeOrderAlert } from "./ItemOpenChangeOrderAlert";
