--- apps/erp/app/modules/items/ui/ChangeOrderActions/index.ts
@@ -1,2 +1,0 @@
-export { default as ChangeOrderRequiredActionForm } from "./ChangeOrderRequiredActionForm";
-export { default as ChangeOrderRequiredActionsTable } from "./ChangeOrderRequiredActionsTable";
