--- apps/erp/app/modules/items/ui/ChangeOrderTypes/index.ts
@@ -1,2 +1,0 @@
-export { default as ChangeOrderTypeForm } from "./ChangeOrderTypeForm";
-export { default as ChangeOrderTypesTable } from "./ChangeOrderTypesTable";
