+++ apps/erp/app/modules/items/ui/Tools/ToolHeader.tsx
@@ -26,7 +26,7 @@ import ConfirmDelete from "~/components/Modals/ConfirmDelete";
 import { usePermissions, useRouteData, useUser } from "~/hooks";
 import { path } from "~/utils/path";
 import type { Tool } from "../../types";
-import { CreateChangeOrderModal } from "../ChangeOrder";
+import { CreateChangeOrderModal } from "../ChangeNotice";
 import { getItemLifecycleStatus } from "../Item/ItemSupersessionForm";
 import { useToolNavigation } from "./useToolNavigation";
