+++ apps/erp/app/routes/x+/items+/change-notice-actions.$id.tsx
@@ -9,7 +9,7 @@ import {
   getChangeOrderRequiredAction,
   upsertChangeOrderRequiredAction
 } from "~/modules/items";
-import { ChangeOrderRequiredActionForm } from "~/modules/items/ui/ChangeOrderActions";
+import { ChangeOrderRequiredActionForm } from "~/modules/items/ui/ChangeNoticeActions";
 import { path } from "~/utils/path";
 
 export async function loader({ request, params }: LoaderFunctionArgs) {
