+++ apps/erp/app/routes/x+/tool+/$itemId.details.tsx
@@ -29,7 +29,7 @@ import { getRevisionLock } from "~/modules/items/items.server";
 import {
   ItemChangeOrders,
   ItemOpenChangeOrderAlert
-} from "~/modules/items/ui/ChangeOrder";
+} from "~/modules/items/ui/ChangeNotice";
 import {
   BillOfMaterial,
   BillOfProcess,
